requirements.txt
torch
torchvision
numpy
pyyaml
pillow
tqdm
matplotlib
tensorboard
lpips
thop
tabulate
pytest
